refactor: split main.py into startup, lifecycle, and router modules
Extract early-boot side effects (env vars, PROJ config, logging, dotenv,
OpenAPI generation) into startup.py, Prefect lifespan management into
lifecycle.py, and the /ogcapi redirect into routers/root.py. Reduces
main.py from 138 lines to ~25 lines of pure app wiring.
